backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "celery>=5.6.3",
    "email-validator>=2.3.0",
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "openai>=2.33.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.14.0",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.27",
    "redis>=7.4.0",
    "sqlalchemy[asyncio]>=2.0.49",
    "uvicorn[standard]>=0.46.0",
]
